Fix performance

+++ Tibber.Robot.Api/Models/Execution.cs
@@ -5,6 +5,6 @@ public class Execution
     public int Id { get; set; }
     public required DateTimeOffset Timestamp { get; set; }
     public required int Commands { get; set; }
-    public required int Result { get; set; }
+    public required long Result { get; set; }
     public required TimeSpan Duration { get; set; }
 }
